# openapi.yaml
openapi: 3.0.3
info:
  title: Наш дом API
  description: |
    REST API мини-приложения «Наш дом» для мессенджера MAX.
    Базовый префикс всех методов: `/server`.
  version: 1.0.0

servers:
  - url: https://{domain}
    description: Продакшен (Caddy HTTPS)
    variables:
      domain:
        default: example.com
  - url: http://127.0.0.1:3080
    description: Локальный доступ к контейнеру bot

tags:
  - name: system
  - name: auth
  - name: me
  - name: houses
  - name: chats
  - name: topics
  - name: tickets
  - name: works

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session-token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Status:
      type: object
      required: [ok, service, status]
      properties:
        ok:
          type: boolean
          example: true
        service:
          type: string
          example: nash-dom-api
        status:
          type: string
          example: up
        time:
          type: string
          format: date-time
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [resident, uk, admin]
        houseId:
          type: string
          nullable: true
    LoginRequest:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
          example: "+79001234567"
        password:
          type: string
          example: "1234"
    Bootstrap:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        houses:
          type: array
          items:
            type: object
        chats:
          type: array
          items:
            type: object
        messages:
          type: array
          items:
            type: object
        topics:
          type: array
          items:
            type: object
        tickets:
          type: array
          items:
            type: object
        works:
          type: array
          items:
            type: object

paths:
  /server/status:
    get:
      tags: [system]
      summary: Статус API
      operationId: getStatus
      responses:
        "200":
          description: Сервис доступен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Status"

  /server/health:
    get:
      tags: [system]
      summary: Health-check
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [ok, service]
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string

  /server/openapi.yaml:
    get:
      tags: [system]
      summary: OpenAPI-спецификация
      responses:
        "200":
          description: YAML OpenAPI 3.0
          content:
            text/yaml:
              schema:
                type: string

  /server/DATA-API.yaml:
    get:
      tags: [system]
      summary: Конфиг обязательных проверок API
      responses:
        "200":
          description: DATA-API.yaml
          content:
            text/yaml:
              schema:
                type: string

  /server/auth/login:
    post:
      tags: [auth]
      summary: Вход по телефону и паролю
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Успешный вход (аккаунт создаётся автоматически при первом входе)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
        "401":
          description: Неверный пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /server/auth/register:
    post:
      tags: [auth]
      summary: Регистрация
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone, password, role]
              properties:
                name:
                  type: string
                phone:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [resident, uk]
                ukName:
                  type: string
      responses:
        "200":
          description: Зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
        "409":
          description: Телефон уже занят

  /server/auth/logout:
    post:
      tags: [auth]
      summary: Выход
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean

  /server/auth/max:
    post:
      tags: [auth]
      summary: Вход через MAX Bridge (initData)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [initData]
              properties:
                initData:
                  type: string
                platform:
                  type: string
      responses:
        "200":
          description: Сессия создана
        "401":
          description: Невалидный initData

  /server/me:
    get:
      tags: [me]
      summary: Текущий пользователь и bootstrap-данные
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bootstrap"
        "401":
          description: Нет сессии

  /server/houses:
    get:
      tags: [houses]
      summary: Список домов
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /server/me/house:
    post:
      tags: [me]
      summary: Выбрать дом
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [houseId]
              properties:
                houseId:
                  type: string
      responses:
        "200":
          description: OK

  /server/me/address:
    post:
      tags: [me]
      summary: Сохранить или пропустить адрес
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                street:
                  type: string
                entrance:
                  type: string
                flat:
                  type: string
                skip:
                  type: boolean
      responses:
        "200":
          description: OK

  /server/chats:
    get:
      tags: [chats]
      summary: Чаты дома
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /server/chats/{chatId}/messages:
    get:
      tags: [chats]
      summary: Сообщения чата
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    post:
      tags: [chats]
      summary: Отправить сообщение
      security:
        - bearerAuth: []
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
      responses:
        "200":
          description: Отправлено
        "422":
          description: Отклонено модерацией

  /server/topics:
    get:
      tags: [topics]
      summary: Темы дома
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
    post:
      tags: [topics]
      summary: Создать тему
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [category, title, description]
              properties:
                category:
                  type: string
                title:
                  type: string
                description:
                  type: string
      responses:
        "201":
          description: Создано

  /server/tickets:
    get:
      tags: [tickets]
      summary: Заявки пользователя / УК
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /server/tickets/{id}:
    patch:
      tags: [tickets]
      summary: Сменить статус заявки (УК)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        "200":
          description: OK

  /server/works:
    get:
      tags: [works]
      summary: Работы по подъездам
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK

  /server/works/{id}:
    patch:
      tags: [works]
      summary: Сменить статус работы (УК)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
      responses:
        "200":
          description: OK